+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Timestamp support with `xan plot -LT`.
 * Adding `xan rename -n/--no-headers` support for `-p/--prefix` & `-x/--suffix`.
 * Adding `xan from -f parquet` (requires the `parquet` feature).
+* Adding `range` moonblade function.
 
 ## 0.56.0